[AI] [SPRINT-1] Add self-improving SKILL.md system with auto-creation and curator
