+++ src/agent-runs/api.ts
@@ -279,8 +279,8 @@ export async function agentRunsRoutes(fastify: FastifyInstance, options?: AgentR
 
   // GET /agent-runs/blacklist — list blacklisted issues
   fastify.get("/agent-runs/blacklist", async (request, reply) => {
-    if (!request.userId) {
-      return reply.code(401).send({ error: "Authentication required" });
+    if (!authenticateWrite(request)) {
+      return reply.code(401).send({ error: "Invalid API key" });
     }
     const query = request.query as { workspace?: string };
     const items = db.getBlacklistedIssues(query.workspace);
